+++ app/buttons/recruit/components/create_recruit_buttons.js
@@ -1,6 +1,6 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
 const { URLSearchParams } = require('url');
-const { isNotEmpty } = require('../common');
+const { isNotEmpty } = require('../../../common');
 
 module.exports = {
     recruitDeleteButton: recruitDeleteButton,
