+++ app/cmd/splat3/recruit/salmon_recruit.js
@@ -1,7 +1,11 @@
 const { searchMessageById } = require('../../../manager/messageManager');
 const { searchMemberById } = require('../../../manager/memberManager');
 const { isNotEmpty, sleep } = require('../../../common');
-const { recruitActionRow, recruitDeleteButton, unlockChannelButton } = require('../../../buttons/create_recruit_buttons');
+const {
+    recruitActionRow,
+    recruitDeleteButton,
+    unlockChannelButton,
+} = require('../../../buttons/recruit/components/create_recruit_buttons');
 const { AttachmentBuilder, PermissionsBitField } = require('discord.js');
 const log4js = require('log4js');
 const { fetchSchedule, checkBigRun } = require('../../../common/apis/splatoon3_ink');
