+++ app/modals/recruit/event/regular_recruit_modal.js
@@ -1,8 +1,8 @@
 const RecruitService = require('../../../../db/recruit_service');
-const { getMemberMentions } = require('../../../event/recruit_button');
+const { getMemberMentions } = require('../../../buttons/recruit/event/recruit_button_events');
 const { searchMessageById } = require('../../../manager/messageManager');
 const { isNotEmpty, isEmpty, sleep } = require('../../../common');
-const { recruitActionRow, recruitDeleteButton } = require('../../../buttons/create_recruit_buttons');
+const { recruitActionRow, recruitDeleteButton } = require('../../../buttons/recruit/components/create_recruit_buttons');
 const { setButtonDisable } = require('../../../common/button_components');
 const { AttachmentBuilder } = require('discord.js');
 const { recruitRegularCanvas, ruleRegularCanvas } = require('../../../canvas/recruit/regular_canvas');
